+++ spec/rubocop/cop/rspec/receive_counts_spec.rb
@@ -78,6 +78,13 @@ RSpec.describe RuboCop::Cop::RSpec::ReceiveCounts do
     RUBY
   end
 
+  it 'flags exactly(1).times when not called on `receive`' do
+    expect_offense(<<-RUBY)
+      expect(action).to have_published_event.exactly(2).times
+                                            ^^^^^^^^^^^^^^^^^ Use `.twice` instead of `.exactly(2).times`.
+    RUBY
+  end
+
   include_examples 'autocorrect',
                    'expect(foo).to receive(:bar).exactly(1).times { true }',
                    'expect(foo).to receive(:bar).once { true }'
@@ -85,4 +92,9 @@ RSpec.describe RuboCop::Cop::RSpec::ReceiveCounts do
   include_examples 'autocorrect',
                    'expect(foo).to receive(:bar).at_least(2).times { true }',
                    'expect(foo).to receive(:bar).at_least(:twice) { true }'
+
+  # Does not auto-correct if not part of the RSpec API
+  include_examples 'autocorrect',
+                   'expect(foo).to have_published_event(:bar).exactly(2).times',
+                   'expect(foo).to have_published_event(:bar).exactly(2).times'
 end